refactor: Rename EngineServiceImpl.{h, cpp} to engineService.{h, cpp} to be inline with naming convetions
